+++ msg/CameraCmd.msg
@@ -1,2 +1,2 @@
 int8 device # integer corresponding to X in /dev/videoX
-int8 resolution # 0 is lowest, 1 is medium, 2 is highest
+int8 resolution # 0-4
